+++ previous-versions/common/RELEASES.md
@@ -641,3 +641,13 @@ Based on:
 - [typescript v0.36.0] previous-versions/common
 ### Releases
 - [NPM v0.36.0] https://www.npmjs.com/package/@codat/common/v/0.36.0 - previous-versions/common
+
+## 2023-09-07 00:03:42
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.0] previous-versions/common
+### Releases
+- [NPM v0.37.0] https://www.npmjs.com/package/@codat/common/v/0.37.0 - previous-versions/common